lib/sql: fix syntax in PostgreSQL "grant-all" script
When table names were sorted alphabetically in the "grant-all" script
for PostgreSQL, commas were left unchanged, creating syntax errors in
the commands: some names have no comma to separate them from the
following names, and the last names of the blocks are now followed by a
comma when they should not.

Address it by adding or removing commas in the script.

Fixes: ca0e79d4db34 ("sql: Sort 'grant-all' scripts alphabetically")
Signed-off-by: Quentin Monnet <quentin.monnet@netronome.com>